GUI-App/requirements.txt
wheel
pyside2
gphoto2
opencv-python-headless
pyserial
shortuuid
qrcode
pillow